requirements.txt
slack-sdk
python-dotenv
argparse